+++ src/degenbot/uniswap/v3_libraries/tick.py
@@ -3,7 +3,6 @@ from decimal import Decimal
 from ...constants import MAX_UINT128
 from ...exceptions import EVMRevertError
 from . import tick_math as TickMath
-from .functions import uint24
 
 # type hinting aliases
 Int24 = int
@@ -15,7 +14,7 @@ def tickSpacingToMaxLiquidityPerTick(tickSpacing: Int24):
 
     minTick = Decimal(TickMath.MIN_TICK) // tickSpacing * tickSpacing
     maxTick = Decimal(TickMath.MAX_TICK) // tickSpacing * tickSpacing
-    numTicks = uint24((maxTick - minTick) // tickSpacing) + 1
+    numTicks = ((maxTick - minTick) // tickSpacing) + 1
 
     result = MAX_UINT128 // numTicks
